+++ sys/net/gnrc/routing/rpl/gnrc_rpl_internal/netstats.h
@@ -25,9 +25,6 @@ extern "C" {
 
 #include "net/rpl/rpl_netstats.h"
 
-#define ENABLE_DEBUG    (0)
-#include "debug.h"
-
 #define GNRC_RPL_NETSTATS_MULTICAST (0)
 #define GNRC_RPL_NETSTATS_UNICAST   (1)
